added google pay and apple pay, restyled payment step on checkout
