--- packages/commerce/components/checkout/contact-info.tsx
@@ -1,78 +1,0 @@
-'use client'
- 
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from '@hanzo/ui/primitives/form'
-import { Input } from '@hanzo/ui/primitives'
-import type { UseFormReturn } from 'react-hook-form'
- 
-const ContactInfo: React.FC<{
-  form: UseFormReturn<{
-    firstName: string
-    lastName: string
-    email: string
-  }, any, {
-    firstName: string
-    lastName: string
-    email: string
-  }>,
-}> = ({
-  form,
-}) => {
-  return (
-    <Form {...form}>
-      <form className='text-left'>
-        <div className='flex flex-col gap-4'>
-          <div className='flex flex-col sm:flex-row gap-4'>
-            <FormField
-              control={form.control}
-              name='firstName'
-              render={({ field }) => (
-                <FormItem className='space-y-1 w-full'>
-                  <FormLabel>First name</FormLabel>
-                  <FormControl>
-                    <Input {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-            <FormField
-              control={form.control}
-              name='lastName'
-              render={({ field }) => (
-                <FormItem className='space-y-1 w-full'>
-                  <FormLabel>Last name</FormLabel>
-                  <FormControl>
-                    <Input {...field} />
-                  </FormControl>
-                  <FormMessage />
-                </FormItem>
-              )}
-            />
-          </div>
-          <FormField
-            control={form.control}
-            name='email'
-            render={({ field }) => (
-              <FormItem className='space-y-1 w-full'>
-                <FormLabel>Email</FormLabel>
-                <FormControl>
-                  <Input {...field} />
-                </FormControl>
-                <FormMessage />
-              </FormItem>
-            )}
-          />
-        </div>
-      </form>
-    </Form>
-  )
-}
-
-export default ContactInfo
